+++ packages/yourloops/lib/notifications/notification.hook.tsx
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021-2022, Diabeloop
+ * Copyright (c) 2021-2023, Diabeloop
  *
  * All rights reserved.
  *
@@ -54,12 +54,12 @@ function NotificationContextImpl(): NotificationContext {
   }
 
   const accept = async (notification: Notification): Promise<void> => {
-    log.info('Accept invitation', notification)
+    log.info('Accept invite', notification)
     await NotificationApi.acceptInvitation(user.id, notification)
   }
 
   const decline = async (notification: Notification): Promise<void> => {
-    log.info('Decline invitation', notification)
+    log.info('Decline invite', notification)
     await NotificationApi.declineInvitation(user.id, notification)
     const r = await NotificationApi.getReceivedInvitations(user.id)
     setReceivedInvitations(r)
@@ -79,10 +79,10 @@ function NotificationContextImpl(): NotificationContext {
     await NotificationApi.cancelRemoteMonitoringInvite(teamId, userId)
   }
 
-  const getInvitation = (teamId: string): Notification => {
-    const invitation = sentInvitations.find(invitation => invitation.target.id === teamId)
+  const getInvitation = (teamId: string, patientEmail: string): Notification => {
+    const invitation = sentInvitations.find(invitation => invitation.target.id === teamId && invitation.email === patientEmail)
     if (!invitation) {
-      throw Error(`Could not find invitation for team ${teamId}`)
+      throw Error(`Could not find invite for team ${teamId}`)
     }
     return invitation
   }