+++ packages/yourloops/test/integration/assert/add-patient.assert.ts
@@ -39,7 +39,7 @@ export const checkAddPatientPrivateButtonTooltip = async () => {
 
   const addPatientHoverZone = within(patientListHeader).getByTestId('add-patient-button')
   await userEvent.hover(addPatientHoverZone)
-  const informationTooltip = screen.getByText('To invite a patient, you must first select a care team from the dropdown menu. You can create you own care team if you need to. Alternatively, you can provide the patient with your YourLoops email address so they can enable private data sharing with you.')
+  const informationTooltip = await screen.findByText('To invite a patient, you must first select a care team from the dropdown menu. You can create you own care team if you need to. Alternatively, you can provide the patient with your YourLoops email address so they can enable private data sharing with you.')
   expect(informationTooltip).toBeVisible()
 
   await userEvent.unhover(addPatientHoverZone)
@@ -117,7 +117,7 @@ export const checkAddPatientMedicalTeamDialogInvite = async () => {
   expect(PatientApi.invitePatient).toHaveBeenCalledWith({ teamId: myThirdTeamId, email: 'new-patient@email.com' })
   const identificationCodeDialog = screen.getByRole('dialog')
   expect(identificationCodeDialog).toHaveTextContent('A - MyThirdTeam - to be deletedCommunicate this identification code to your patient during a consultation so they can verify your identity.')
-  expect(identificationCodeDialog).toHaveTextContent('This identification code is always available in the team information on the "My teams" page.263 - 381 - 988Ok')
+  expect(identificationCodeDialog).toHaveTextContent('This identification code is always available in the Care team page.263 - 381 - 988Ok')
 
   const okButton = within(identificationCodeDialog).getByRole('button', { name: 'Ok' })