+++ packages/yourloops/test/integration/mock/team.api.mock.ts
@@ -55,7 +55,7 @@ export const monitoringAlertsParameters: MonitoringAlertsParameters = {
   nonDataTxThreshold: 15,
   reportingPeriod: 7
 }
-export const monitoringAlertsParametersBgUnitMmol = {
+export const monitoringAlertsParametersBgUnitMmol: MonitoringAlertsParameters = {
   bgUnit: Unit.MmolPerLiter,
   lowBg: 2.8,
   highBg: 7.8,
@@ -66,19 +66,22 @@ export const monitoringAlertsParametersBgUnitMmol = {
   reportingPeriod: 7
 }
 
-export const myTeamId = 'myTeamId'
+export const myFirstTeamId = 'myTeamId'
 export const mySecondTeamId = 'mySecondTeamId'
 export const myThirdTeamId = 'myThirdTeamId'
 export const filtersTeamId = 'filtersTeamId'
 export const myFirstTeamName = 'MyFirstTeam'
 export const mySecondTeamName = 'MySecondTeam'
 export const myThirdTeamName = 'A - MyThirdTeam - to be deleted'
+export const myThirdTeamPhoneNumber = '0476000000'
+export const myThirdTeamCode = '263381988'
+export const myThirdTeamAddress = 'Guinness Road W1D 1BS Dublin GB'
 export const filtersTeamName = 'Team used to test the patients filters'
 
 export const buildTeamOne = (): Team => {
   return {
     name: myFirstTeamName,
-    id: myTeamId,
+    id: myFirstTeamId,
     code: '036038775',
     type: TeamType.medical,
     phone: '0476000000',
@@ -138,7 +141,7 @@ export const buildTeamThree = (): Team => {
     type: TeamType.medical,
     phone: '0476000000',
     email: 'hey@third.hey',
-    address: { line1: 'Guiness Road', line2: '', zip: 'W1D 1BS', city: 'Dublin', country: 'IE' },
+    address: { line1: 'Guinness Road', line2: '', zip: 'W1D 1BS', city: 'Dublin', country: 'GB' },
     members: [{
       userId: loggedInUserId,
       role: TeamMemberRole.admin,
@@ -225,7 +228,8 @@ export const iTeamOne: ITeam = {
     city: 'Paris',
     country: 'FR'
   },
-  members: null
+  members: null,
+  monitoringAlertsParameters
 }
 
 export const buildAvailableTeams = (): Team[] => {