+++ packages/yourloops/test/integration/use-cases/patient-list-management.ts
@@ -37,8 +37,10 @@ import {
   checkPatientListPendingTab,
   checkPatientListTooltipsMgDL,
   checkPatientListTooltipsMmolL,
+  checkPendingPatientColumnsSettingsMedicalTeam,
   type Router
 } from '../assert/patient-list.assert'
+import { checkPatientFiltersForPrivateTeam } from '../assert/patient-filters.assert'
 
 export const testPatientListForHcp = async (router: Router) => {
   checkPatientListHeaderForHcp()
@@ -47,6 +49,7 @@ export const testPatientListForHcp = async (router: Router) => {
   await checkMonitoringAlertsIconsInactiveForFirstPatient()
   await checkPatientListFilters()
   await checkPatientColumnsFiltersContent()
+  await checkPendingPatientColumnsSettingsMedicalTeam()
   await checkPatientListHideShowColumns()
   await checkPatientListPendingTab(router)
   await checkPatientListCurrentTab()
@@ -58,4 +61,5 @@ export const testPatientListForHcpWithMmolL = async () => {
 
 export const testPatientListForHcpPrivateTeam = async () => {
   await checkPatientListCurrentTabForPrivateTeam()
+  await checkPatientFiltersForPrivateTeam()
 }