+++ packages/yourloops/test/unit/pages/hcp/patients/patient-utils.test.tsx
@@ -31,6 +31,7 @@ import { UserInvitationStatus } from '../../../../../lib/team/models/enums/user-
 import { TeamMemberRole } from '../../../../../lib/team/models/enums/team-member-role.enum'
 import { type ITeamMember } from '../../../../../lib/team/models/i-team-member.model'
 import { type Profile } from '../../../../../lib/auth/models/profile.model'
+import { Gender } from '../../../../../lib/auth/models/enums/gender.enum'
 
 describe('Patient utils', () => {
   describe('mapTeamUserToPatient', () => {
@@ -58,11 +59,18 @@ describe('Patient utils', () => {
           nonDataTransmissionRate: 10,
           nonDataTransmissionActive: true
         },
+        glycemiaIndicators: {
+          timeInRange: 3,
+          hypoglycemia: 20,
+          coefficientOfVariation: 1,
+          glucoseManagementIndicator: 0
+        },
         unreadMessages: 5
       }
       teamMember.unreadMessages = 4
       const patient: Patient = {
         monitoringAlerts: teamMember.alarms,
+        glycemiaIndicators: teamMember.glycemiaIndicators,
         metadata: {
           flagged: undefined,
           medicalData: null,
@@ -75,7 +83,7 @@ describe('Patient utils', () => {
           lastName: profile.lastName,
           email: teamMember.email,
           referringDoctor: undefined,
-          sex: ''
+          sex: Gender.NotDefined
         },
         settings: {
           a1c: undefined,