+++ CHANGELOG.md
@@ -29,10 +29,15 @@ It is based on Tidepool Blip 1.27.
 - YLP-2220 Remove "private practice" filter from patients list
 - YLP-2231 Implement new medical reports design for patients
 - YLP-2265 Disable Add patient button when private practice is selected
+- YLP-2285 Monitoring filter cannot be enabled when Private practice is selected
 - YLP-2312 Teams and patients in a team should always have a field "monitoringAlertsParameters"
 - YLP-2322 Pending patients should not display glycemia specific data
 - YLP-2323 Add new columns from existing data in patient list
 - YLP-2324 Hide certain columns when private practice is selected
+- YLP-2348 New columns from glycemia indicators in patients list
+- YLP-2350 Display a visual info when a patient has unread messages from another team in the chat widget
+- YLP-2358 As a care team member, I want a new design when listing my pending patients
+- YLP-2390 Enlarge font of the buttons of the messaging system
 
 ### Engineering use
 - YLP-2044 Remove redux settings store
@@ -82,9 +87,14 @@ It is based on Tidepool Blip 1.27.
 - YLP-2293 Migrate Average Glucose from stat.js to yourloops
 - YLP-2300 Remove the remote monitoring invitation flow
 - YLP-2301 Split ITs jobs by roles
+- YLP-2307 Incomplete alert when accepting a direct share invitation
 - YLP-2311 Reorganize ITs
 - YLP-2351 migrate average daily dose stat and total insulin stat from stat.js to yourloops
+- YLP-2364 Fixed English translations
 - YLP-2367 Patient list refacto
+- YLP-2386 Fix navigation crash between care team and patient list
+- YLP-2387 Fix the tab title when displaying the patient list
+- YLP-2404 Fix patient list columns width
 
 ### Fixed
 - YLP-2145 Patient cannot see monitoring widgets when he's in a monitoring team and another non-monitoring team